requirements-minimal.txt
# Minimal requirements for Render deployment
# Core functionality only for faster builds

# Web framework
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# Core LlamaIndex
llama-index-core>=0.14.0
llama-index-llms-gemini>=0.6.0
llama-index-embeddings-huggingface>=0.6.0
llama-index-vector-stores-qdrant>=0.8.0

# Essential ML packages
torch>=2.0.0
transformers>=4.30.0,<5.0.0
sentence-transformers

# Vector database
qdrant-client

# Document processing
pymupdf4llm

# Utilities
python-dotenv
numpy
pandas
requests
nltk

# Google API
google-generativeai